chore(nvidia_transformer_xl): First cleanup to only add search-related functions and files.
